requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
pillow
pdfplumber
pytesseract
celery
redis
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
pillow
pdfplumber
pytesseract
celery
redis